+++ api/os/os.cpp
@@ -87,6 +87,32 @@ void __dispatchSpawnedProcessEvt(int virtualPid, const string &action, const jso
     events::dispatch("spawnedProcess", evt);
 }
 
+os::LocaleInfo __parseLocaleInfo(string localeName) {
+    size_t encodingIndex = localeName.find('.');
+    if(encodingIndex != string::npos) {
+        localeName = localeName.substr(0, encodingIndex);
+    }
+
+    size_t modifierIndex = localeName.find('@');
+    if(modifierIndex != string::npos) {
+        localeName = localeName.substr(0, modifierIndex);
+    }
+
+    replace(localeName.begin(), localeName.end(), '_', '-');
+
+    os::LocaleInfo localeInfo;
+    localeInfo.locale = localeName;
+
+    vector<string> localeParts = helpers::split(localeName, '-');
+    if(localeParts.size() > 0) {
+        localeInfo.language = localeParts[0];
+    }
+    if(localeParts.size() > 1) {
+        localeInfo.region = localeParts[localeParts.size() - 1];
+    }
+    return localeInfo;
+}
+
 bool isTrayInitialized() {
     return trayInitialized;
 }
@@ -259,9 +285,8 @@ bool updateSpawnedProcess(const os::SpawnedProcessEvent &evt) {
     return true;
 }
 
-string __getLocaleName() {
+os::LocaleInfo getLocaleInfo() {
     string locale;
-
     #if defined(_WIN32)
     wchar_t localeName[LOCALE_NAME_MAX_LENGTH] = {0};
     if(GetUserDefaultLocaleName(localeName, LOCALE_NAME_MAX_LENGTH) > 0) {
@@ -296,38 +321,8 @@ string __getLocaleName() {
         const char *currentLocale = setlocale(LC_ALL, nullptr);
         locale = currentLocale == nullptr ? "" : string(currentLocale);
     }
-    return locale;
-}
 
-os::LocaleInfo __parseLocaleInfo(string localeName) {
-    size_t encodingIndex = localeName.find('.');
-    if(encodingIndex != string::npos) {
-        localeName = localeName.substr(0, encodingIndex);
-    }
-
-    size_t modifierIndex = localeName.find('@');
-    if(modifierIndex != string::npos) {
-        localeName = localeName.substr(0, modifierIndex);
-    }
-
-    replace(localeName.begin(), localeName.end(), '_', '-');
-
-    os::LocaleInfo localeInfo;
-    localeInfo.locale = localeName;
-
-    vector<string> localeParts = helpers::split(localeName, '-');
-    if(localeParts.size() > 0) {
-        localeInfo.language = localeParts[0];
-    }
-    if(localeParts.size() > 1) {
-        localeInfo.region = localeParts[localeParts.size() - 1];
-    }
-
-    return localeInfo;
-}
-
-os::LocaleInfo getLocale() {
-    return __parseLocaleInfo(__getLocaleName());
+    return __parseLocaleInfo(locale);
 }
 
 string getPath(const string &name) {
@@ -382,6 +377,16 @@ string getEnv(const string &key) {
     char *value;
     value = getenv(key.c_str());
     return value == nullptr ? "" : string(value);
+    #endif
+}
+
+bool setEnv(const string &key, const string &value) {
+    #if defined(_WIN32)
+    wstring wideKey = helpers::str2wstr(key);
+    wstring wideValue = helpers::str2wstr(value);
+    return _wputenv_s(wideKey.c_str(), wideValue.c_str()) == 0;
+    #else
+    return setenv(key.c_str(), value.c_str(), 1) == 0;
     #endif
 }
 
@@ -528,6 +533,25 @@ json getEnv(const json &input) {
     return output;
 }
 
+json setEnv(const json &input) {
+    json output;
+    const auto missingRequiredField = helpers::missingRequiredField(input, {"key", "value"});
+    if(missingRequiredField) {
+        output["error"] = errors::makeMissingArgErrorPayload(missingRequiredField.value());
+        return output;
+    }
+    string key = input["key"].get<string>();
+    string value = input["value"].get<string>();
+
+    if(os::setEnv(key, value)) {
+        output["success"] = true;
+    }
+    else {
+        output["error"] = errors::makeErrorPayload(errors::NE_OS_UNLTOUV);
+    }
+    return output;
+}
+
 json getEnvs(const json &input) {
     json output;
     #if defined(__linux__) || defined(__FreeBSD__) || defined(__APPLE__)
@@ -562,9 +586,9 @@ json getEnvs(const json &input) {
     return output;
 }
 
-json getLocale(const json &input) {
+json getLocaleInfo(const json &input) {
     json output;
-    os::LocaleInfo localeInfo = os::getLocale();
+    os::LocaleInfo localeInfo = os::getLocaleInfo();
 
     json retVal;
     retVal["locale"] = localeInfo.locale;