+++ api/os/os.h
@@ -47,8 +47,9 @@ os::CommandResult execCommand(string command, const os::ChildProcessOptions &opt
 pair<int, int> spawnProcess(string command, const os::ChildProcessOptions &options = {});
 bool updateSpawnedProcess(const os::SpawnedProcessEvent &evt);
 string getPath(const string &name);
-os::LocaleInfo getLocale();
+os::LocaleInfo getLocaleInfo();
 string getEnv(const string &key);
+bool setEnv(const string &key, const string &value);
 bool trashItem(const string &path);
 
 namespace controllers {
@@ -58,8 +59,9 @@ json spawnProcess(const json &input);
 json updateSpawnedProcess(const json &input);
 json getSpawnedProcesses(const json &input);
 json getEnv(const json &input);
+json setEnv(const json &input);
 json getEnvs(const json &input);
-json getLocale(const json &input);
+json getLocaleInfo(const json &input);
 json showOpenDialog(const json &input);
 json showFolderDialog(const json &input);
 json showSaveDialog(const json &input);